feat: overrideStyles added to useTheme context, extracts useColorOverrides hook
