+++ lib/components/OverdriveProvider/OverdriveProvider.tsx
@@ -4,22 +4,17 @@ import { isEqual } from 'es-toolkit';
 import React, { createContext, useContext, useEffect, useMemo } from 'react';
 
 import baseTheme from '../../themes/base';
-import { shadedColour } from '../../themes/helpers';
 import { makeRuntimeTokens, RuntimeTokens } from '../../themes/makeTheme';
 import { themeContractVars } from '../../themes/theme.css';
 import { BreakPoints } from '../../themes/tokens';
 import { isBrowser } from '../../utils';
 
-type ThemeContract = typeof themeContractVars;
-export interface ThemeOverrides {
-	primaryColourBackground: string;
-	primaryColourBackgroundMild: string;
-	primaryColourBackgroundStrong: string;
-	primaryColourBorder: string;
-	primaryColourForeground: string;
-}
+import { useColorOverrides, type ColorOverrides } from './useColorOverrides';
 
-export interface BaseProps {
+type ThemeContract = typeof themeContractVars;
+type PortalMountPoint = React.RefObject<HTMLElement | null>;
+
+export interface ProviderProps {
 	/** The theme object to be used throughout the application */
 	theme?: typeof baseTheme;
 	/** Custom breakpoints for responsive design */
@@ -27,21 +22,22 @@ export interface BaseProps {
 	/** When true, prevents applying theme styles at the body level */
 	noBodyLevelTheming?: boolean;
 	/** Custom colour overrides for select theme tokens */
-	overrides?: Partial<ThemeOverrides>;
+	overrides?: Partial<ColorOverrides>;
 	/** Reference to an HTML element where portals should be mounted */
-	portalMountPoint?: React.RefObject<HTMLElement | null>;
+	portalMountPoint?: PortalMountPoint;
+	children?: React.ReactNode;
 }
 
-export interface ProviderContext extends BaseProps {
+export interface ProviderContext
+	extends Pick<ProviderProps, 'portalMountPoint'> {
+	themeClass: (typeof baseTheme)['themeRef'];
+	overrideStyles: ReturnType<typeof assignInlineVars>;
 	vars: ThemeContract;
-	themeClass: string;
 }
-
-export type ProviderProps = React.PropsWithChildren<BaseProps>;
 
 const OverdriveContext = createContext<ProviderContext | null>(null);
 const RuntimeTokensContext = createContext<RuntimeTokens | null>(null);
-const msgInvariantError = "You haven't provided an `OverdriveProvider`.";
+const msgInvariantError = "You haven't used an `<OverdriveProvider>`.";
 
 export const useTheme = () => {
 	const context = useContext(OverdriveContext);
@@ -55,97 +51,12 @@ export const useRuntimeTokens = (): RuntimeTokens => {
 	return tokens;
 };
 
-const isValidColor = (color: string): boolean => {
-	const s = new Option().style;
-	s.color = color;
-	return s.color !== '';
-};
-
 const isSafeElement = (element: HTMLElement | null): boolean => {
 	return (
 		element instanceof HTMLElement &&
 		!element.hasAttribute('onclick') &&
 		!element.hasAttribute('onerror')
 	);
-};
-
-const useOverrides = (
-	overrides: Partial<ThemeOverrides> | undefined,
-	themeMode: string,
-) => {
-	return useMemo(() => {
-		if (!overrides) return {};
-
-		// Validate all color inputs
-		const colorKeys = [
-			'primaryColourBackground',
-			'primaryColourBackgroundMild',
-			'primaryColourBackgroundStrong',
-			'primaryColourBorder',
-			'primaryColourForeground',
-		] as const;
-
-		colorKeys.forEach((key) => {
-			if (overrides[key] && !isValidColor(overrides[key]!)) {
-				console.warn(
-					`Invalid color value for ${key}: ${overrides[key]}`,
-				);
-				delete overrides[key];
-			}
-		});
-
-		let mildPrimary: string | null = null;
-		let strongPrimary: string | null = null;
-
-		if (overrides.primaryColourBackground) {
-			mildPrimary =
-				overrides.primaryColourBackgroundMild ||
-				shadedColour({
-					colour: overrides.primaryColourBackground,
-					isDarkTheme: false,
-					direction: themeMode === 'light' ? 'forward' : 'backward',
-					intensity: 0.1,
-				});
-			strongPrimary =
-				overrides.primaryColourBackgroundStrong ||
-				shadedColour({
-					colour: overrides.primaryColourBackground,
-					isDarkTheme: false,
-					direction: themeMode === 'light' ? 'forward' : 'backward',
-					intensity: 0.1,
-				});
-		}
-		// although this look scary, assignInlineVars only generates css vars to apply to a container
-		// anyproperty that is undefined will not have an inline css var generated
-		return assignInlineVars(themeContractVars, {
-			colours: {
-				intent: {
-					primary: {
-						background: {
-							//@ts-expect-error no undefined
-							standard:
-								overrides.primaryColourBackground ?? undefined,
-							//@ts-expect-error no undefined
-							mild: mildPrimary ?? undefined,
-							//@ts-expect-error no undefined
-							strong: strongPrimary ?? undefined,
-						},
-						//@ts-expect-error no undefined
-						foreground:
-							overrides.primaryColourForeground ?? undefined,
-						//@ts-expect-error no undefined
-						border: overrides.primaryColourBorder ?? undefined,
-					},
-				},
-			},
-			typography: {
-				colour: {
-					//@ts-expect-error no undefined
-					primary: overrides.primaryColourBackground ?? undefined,
-				},
-			},
-		});
-	}, [overrides, themeMode]);
 };
 
 export const Provider = ({
@@ -156,21 +67,20 @@ export const Provider = ({
 	portalMountPoint,
 	theme = baseTheme,
 }: ProviderProps) => {
-	const themeValues = useMemo(
-		() => ({
-			vars: theme.vars,
-			themeClass: theme.themeRef,
-			portalMountPoint,
-		}),
-		[portalMountPoint, theme],
-	);
-
 	const runtimeTokens = useMemo(
 		() => makeRuntimeTokens(breakpoints),
 		[breakpoints],
 	);
-
-	const styles = useOverrides(overrides, String(theme.vars.mode));
+	const styles = useColorOverrides(overrides, String(theme.vars.mode));
+	const themeValues = useMemo(
+		() => ({
+			themeClass: theme.themeRef,
+			overrideStyles: styles,
+			portalMountPoint,
+			vars: theme.vars,
+		}),
+		[portalMountPoint, styles, theme],
+	);
 
 	// Body Level Theming
 	useEffect(() => {
